Update service, controller and DTO of "recettes" and "articles" in backend
Updated methods in service and controller of "recettes" and "articles" to work with async/await and promises, added a RecettesCreateDto and utilised it in the service's post methods for creating new "recettes". Renamed "postArticles" and "putArticles" to "postArticle" and "putArticle" in the controller's methods to make it more semantically correct. Implemented TypeORM in "recettes" module and set up a new entity, RecettesEntity, for interacting with the database. Minor changes in frontend to work with the changes in the backend. This update is made to improve code readability and efficiency, and to prepare for database interaction in later features.