"Implement CRUD operations for Recettes with TypeORM"
Updated Recettes service and controller to handle CRUD operations using TypeORM repository. Imported and used necessary decorators and added database queries in the service. Also added NotFoundException handlers. Made similar changes to the articles service and controller as well for consistency. In Recettes entity, changed the name from 'articles' to 'recettes', and added a jsonb column for articleLists. Made a change in the shared-interfaces to ensure the ArticleList interface has the necessary variables required. In the editRecette file in the front application, updated the form to correctly handle addition and deletion of items in an array using yup validation. Finally, updated the Recettes DTO to handle articlesList during creation and updating of a recette.