Add quantity field to calendar items
This change adds a quantity field to the divers items in the calendar. This field allows to ideate multiple quantity of a single item. This was achieved by making some changes in the 'dialog-add-calendar.tsx' and 'diver.dto.ts' of front-app and back-app respectively to accommodate the new quantity field. Also, diversified the entity field in the 'diver.service.ts' to create multiple divers types based on the quantity requested.